+++ feature/history/src/main/java/com/cmc15th/pluv/feature/history/viewmodel/HistoryUiEvent.kt
@@ -2,5 +2,5 @@ package com.cmc15th.pluv.feature.history.viewmodel
 
 sealed class HistoryUiEvent {
     data object OnLoadHistories: HistoryUiEvent()
-    data class OnHistoryClicked(val historyId: Int) : HistoryUiEvent()
+    data class OnHistoryClicked(val historyId: Long) : HistoryUiEvent()
 }
